requirements.txt
# Production requirements - auto-generated from pyproject.toml
# For platforms that don't support pyproject.toml

flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pillow>=10.0.0
pytesseract>=0.3.10
anthropic>=0.25.0
redis>=5.0.0
marshmallow>=3.20.0
werkzeug>=3.0.0
numpy>=2.3.1
scikit-image>=0.25.2
flask-bcrypt>=1.0.1
requests>=2.32.4
gunicorn>=21.2.0
flask-talisman>=1.1.0
flask-limiter>=3.5.0